Shuttle fix, tweak, and verb addition
Stops space tiles from appearing on the sulaco from crash explosions. It is still possible for breaching via destroying the windows and opening doors next to space.
Stops SM explosions from pulling landmarks (tested with the singulo which uses the same pulling proc)
Added a verb to force a dropship launch. Tested with and without crashing, and I did in fact remember a Cancel button this time.