- Adds code to allow certain beds to have bodybags/cryobags buckled to them instead of a mob.
- fire mission flight is renamed 'flyby' flight.

- Adds a dropship medevac equipment that is installed on the MG attach point (at the rear). The system has a winch that is used to lift patient/bodybag/cryobag on a medevac stretcher onto the dropship. The dropship needs to be in flyby flight (formerly 'fire mission' flight) to use the medevac system.
How it works: a combat medic unfolds their medevac stretcher on the planet ground and buckle a marine/bodybag/cryobag to it, then rightclick the stretcher to use the verb that toggles the stretcher's beacon. Then a pilot in flyby flight clicks the medevac system on the dropship equipment console to detect any medevac stretcher beacon. If some are detected, the pilot gets a window to select which stretcher they want to focus on. This 'moves' the dropship over the medevac stretcher, then the pilot must click the medevac system directly with their hand to activate the winch. After a couple seconds the winch lifts what's buckled to the stretcher onto the dropship and instantly brings down the stretcher back on the ground.
If the lift is successful, the medevac system enters a 1 minute cooldown.

The medevac stretcher beacon does not work underground.
The medevac stretcher is a foldable bed that's unmovable while deployed.

Only people with sufficient medical skill can activate the stretcher's beacon.
Only people with pilot skills can use the medevac system's winch.