Improves flaming etcetera. (#341)
...Instead of just doing nothing.
That means you can flame an already flamed turf to replinsh it. Won't stack as it deletes the previous flame.
But also means you can replace a turf's flame with another, be it weaker or stronger... as one would say: "Fight fire with fire".
Kind of also applies to boiler's spray.

Yet another downstream fix...