No scout xeno left behind (#473)
Xenos aboard the Almayer complained they were being deleted despite being forward troops. I've added a Z level check to prevent this (if you're in transit or on the Almayer, you're safe). This doesn't remove the standard check of being with the Queen in case future behaviour involves the Queen being on neither of those Z levels.