G8 Pouch Upgrade
G8 pouch was recently added to surplus clothing vendor and i got chanse to try it out. Turned out that this pouch trades entire suit storage slot for three regular (w_class 3 max) slots. Tweaked it, so it's basically a satchel for suit storage slot now. Controversial, maybe it should be 10, not 15.