Fixes killer ice. (#52)
Current issue with food reagnets is that they affect the body's temperature. There is no cap on this adjustment so if there's a reagent that is very cold like ice (-20) and metabolizes VERY slowly, it will just kill you by dealing burn damage (from being too cold). This puts a cap on the total temperature change allowed and also does the change to that gradually in steps for as long as the reagent is in the body.