Brig/OB room Swap, CIC Update
Swaps the brig and the OB room, with the OB as an extension of the hangar and the Brig in front of CIC. Moves all bunks, including the SO and XO bunks, the CL office and the Officer's mess into the CIC ring. Expands CIC armoury while adding engi/medical equipment, adding a hallway surrounding the CIC itself. Adds white fire extinguishers and megaphones around the map; TCs and POs get their own weapons vendors