Filler Room/area cleanup
Cleaned up a few areas that weren't aligned properly in the brig and the former officer bunks, including the CMP's office and the Cell areas. 
Fixed a literal opening into the void of space in the front of the brig (at the execution chamber), cut off from pasting (my bad).
Adds Experimental (energy) weapon firing range in place of the former SO bunks.
Still need to code separate areas for old bunk areas