Secure Dome, guns
Replaces the bright red floor tiles in secure dome and the corporate dome with more tame tiles, re-does the secure dome's layout to be less like a bunch of random junk scattered on a bunch of shelves and more like a real storage area.
Adds more goodies to the secure dome and Tfort as they are hard to reach and impossible to get to for survivors, respectively (three times as much ammo and another MP5 in Tfort, whilst the secure dome gets 2 MP5's, 2 HG shotguns and 2 MAR carbines. Adds a stack of metal, glass and wood, 3 vests and 3 explosives for survivors to use. Removes flamethrower and pulse rifle in exchange for the more native weapons)
Adds more trees around hydro