Rework of Hit and Run
Hit and run now involves building momentum. For every tile moved, the Hit and Run damage multiplier increases by +5% to a maximum of +100% damage or a 2x damage multiplier. All Hit and Run bonuses are lost if the Runner is stationary for more than 0.5 seconds, and upon hitting a target.