Tank Rebalance
:cl: by Surrealistik
balance: Knockdown stacks for mobs crushed by the tank reduced from 12 to 1, and they are displaced in a random direction. Tanks do not apply knockdown stacks to knocked down mobs. Mobs crushed are displaced in a random direction.
balance: Tank takes no collision damage while running over objects/walls with the snowplow, and the damage it deals to them is doubled.
/:cl: