Tank Rebalance (#1033)
* Tank Rebalance

:cl: by Surrealistik
balance: Knockdown stacks for mobs crushed by the tank reduced from 12 to 1, and they are displaced in a random direction. Tanks do not apply knockdown stacks to knocked down mobs. Mobs crushed are displaced in a random direction.
balance: Tank takes no collision damage while running over objects/walls with the snowplow, and the damage it deals to them is doubled.
/:cl:

* Tweaks

* Tweak

* Requested changes

* Minor fix

* Bug fix for real, fuck tank spaget code

* Tweak to stop benos from getting chained crushed to death by overdrive