Severely nerfs ceremonial sword. (#1158)
Synth was capable of getting 60 + 60% for 104 damage per swing thanks to weapon skills. For mps this is the same, and for specialists this is 82 damage per swing. Captains and other high command too.

People are starting to give the sword to the synth simply because they know synths can hit this much damage.

Machete has 40 damage, significantly lower, no other benefits.

I feel like 60 would've been fine if the captain just uses it for self-defence, but clearly we can't have nice things without abuse...