Minor Ice Update
Adds more to the area between the west underground cargo and LZ1, including the following:

More flanks and rooms at LZ1, including a garage/warehouse west of the landing zone, an extra storage room under Tcomms and a control bubble.
*DS1 Console moved to control bubble

Added lots of fences to the open area between hydro and Dorms, akin to LV624's fencing

Added small extension to tunnel in the middle of the map, small warehouse.

Opened up research with 3 openings in the walls, to serve as an alternate main path