Second batch of /attackby( changes
copypasta:
This will be a long series as there's 330 files where attackby is used which will mean ~8 PRs in total

The goal is for every attackby to call parent where possible for proper fingerprinting and click delays.

No I didn't test this yet so be wary before TMing.

I'm also cleaning up the code SLIGHTLY, changes welcome as long as they're reasonable.
-----
I was much more tired when making this one. Poke me in a reasonable timeframe and ree at me to revisit it
Param -> Params


Compile