Sacrifices sanity to preserve the user experience (#2496)
* Sacrifices sanity to preserve the user experience

Sometimes, writing to a savefile can fail due to it being locked. So far even after a shitton of hours I haven't been able to reproduce when exactly it happens and from speaking with Lummox he said it happens when the savefile is locked. But here's the thing - you can't tell when, why, or because of what it's locked. Nothing in our code uses the same savefile at the same time, so perhaps it may be related to writing to it in sequence often. Whatever the cause is, there's no clean solution here really and I don't want to remove user QoL of every change saving right away just to avoid this bug. It's not pretty but it works.

Fixes: #2281

* h

* e