Prevents pod runtime for admins
if the temp pod gets deleted for some reason(can realistically happen if admins are fucking around) this will close the UI, forcing admins to re-open it to spawn a new one.

The reason I didn't just recreate is so that it gets done properly rather than risking more quirkiness.