Fixes a runtime when clicking with null turf in the clickcatcher (#2599)
The origin turf can be null due to a variety of things, like during the initial mapload if you click fast enough before the titlescreen is loaded. There are some other intended uses, so it's impossible to distinguish whether the current null is intended or not.

This also means the eye check be without the ?, unless I'm missing something.

Anyway, we'll see in testing.