Mass replace MAR40 with AK47
Follows through with request to replace all instances of the MAR40 with the AK47, following the MAR's resprite and rework into the AK47.

Changes to the Heavy Rifle bullet: buffed penetration to 10 from -10, lowered accuracy to -15 from -10, as a bullet known for "Overpenetrating" targets.

Tweaked attachments for the AK 47, can only accept a rail light and a bayonet as a gun made to be "fired from iron sights". Did I say this already? Aligned attachments more properly.