Makee the resin door faster in opening.
Xenos rarely use the resin doors because they take to long to open and marines are only merely more slowed down. The code changes should make them faster. (actually @Hughgent came up with the "changes") I will try to provide a better sprite soon as the current sprite is now way to slow.
An decrease of the time to close is probably not needed.