Makes powergamers sad earlygame
Adds larva jelly to the required ingredients of meralyne and dermaline. Squeezing one larva out gets 50u of Meralyne or Dermaline.

Adds a recipe for Neuraline. Requires one larva to be squeezed as catalyst first, then every other larva souped can make 4u of neuraline when combined with 1u of synaptizine, dylovene, and 2u of tricord. A side effect of the mixing process is the production of 1u of extremely potent neurotoxin which must be filtered out prior to use.