Flamer changes (#4800)
* Fire delay change

Fire delay changed from 20 to 4 (0.4 seconds) for both "full" flamers, should make flamer's more viable as a primary weapon (given that they ARE a primary weapon).

This should keep DPS more-or-less the same, but allow full flamers to compete with miniflamers for area coverage at the cost of drastically upping ammo usage and moderately upping ammo wastage.

* Nerfs both welderpacks

Flamer tanks for the standard flamer hold 75 fuel. A satchel can hold 5 of them, that's 375 fuel. A welding fuel tank (the hueg ones) holds a maximum of 1000u.

Currently the standard welderpack holds 700 fuel (and the engineering welderPACK holds 600 for some weird reason).

I'm dropping both of these down to 500, half-a-welding fuel tank each, for a little under a 50% increase in ammo capacity instead of slightly under 100%.

From personal experience with the miniflamer, it's hard, but doable, to burn through that much fuel in a round. The increased fire rate should make it possible to do so with the regular flamer now. Given as these are available in near-unlimited quantities, I believe it's reasonable.