More Refactors. Disabling Mind Wrack for pending use
Did requested refactors, and improved the weed vibe checks for the Hivemind.

Mind Wrack is disabled for now pending subsequent new features to the Hivemind that may be included in a separate future PR; these will probably include destructible avatars/buildables the Hivemind can make which will be vectors for certain abilities such as Mind Wrack.