Ravager Rework
1. Ravager gains the Rage ability:

Rage
Duration: 10 seconds
Cooldown: 2 Minutes
Effect: The Ravager bellows a demoralizing, earth shaking roar and gains extra speed, armor and slash damage in proportion to the % of its missing HP for the duration. This bonus is further increased for each negative HP the Ravager has. The roar imposes 1-3 slow and stagger stacks (3 * Rage Power) on each living creature within range and earshot (6 * Rage Power), scaling in range and power with Rage strength. 

Rage Power is calculated as 50% of the % difference between the Ravager's current and maximum hit points: (1 - (X / 350) ) * 0.5

So it's 0.5 when the Ravager is at 0 HP, 0.64 when it's at -100 HP and 0.93 when it's at -300 HP.

Rage Power is further increased by 0.005 per point of negative HP, and is capped at 1.

Slash damage is increased by a % = to Rage Power.

Armor is increased by 40 * Rage Power

Speed is increased by 50% of Rage Power.

2. Ignore Pain renamed to Endure, and its duration has been increased to 20 seconds up from 10. It now has an aura SFX and additional notifications.

3. Added a xeno melee damage modifier var, allowing for global modification of slash damage for individual xenos.

4. Added statistics support for the Ravager abilities.