Broken bones do damage
Surrealistic changed bones to do no damage on walk intent, which was good.
Rohesie changed walk intent to be the default, which was also good.

Together, that means that no one takes shrapnel or broken bone damage. This changes that. Also, being pulled (and probably being carried) will be damage-immune.