Fix description
Fixed the description on the Mark 1 Baldur Light Amplification System to match what it does. If you want it to match the rest of the mark ones the slowdown should be set to 0.3 but then literally no one would ever pick it. I just made the description different from the parent to match what it does in game.