Rework some tick defines to better yield to byond's processing needs
This is a copy of an idea I had for /tg/, that doesn't work over there because the mc needed more time then this would allow it to get. Looking at tgmc performance on high pop, this will work over here, as most of the cpu time is spent on byond processing, and most mc usage of cpu time is spikey, meaning it would benefit from automagically spreading across ticks.