Build mode fill tool asks for confirmation if filling area is large (#57007)
Prompts admins to confirm their selection with the build mode fill tool if the selected area is over 150 tiles.

I've seen this happen a few times now (to both myself and others), accidentally filling a massive area because they either forgot that the tool was still on or were moved unexpectedly. This should prevent these kinds of mistakes from happening easily.