If it is free and common, it leaves the medic POINTS vendor.
If it is a common thing available in free vendors, or the medic-only free vendor, it leaves the points vendor.
You do not need to buy a bicardine pill bottle. Just get one.

Bicamera and keloderm pills price lowered to 16 from 30, replaced with meralyne and dermaline pills. No pill mixes roundstart. Price roughly in line with other T2 chems.

Ironsugar case and hypospray replaced with Nanoblood case and spray for same cost.