Duplication fix
Adds a decimal point to fix a duplication issue.

If you have an autolathe and a stack of 20 rglass, you can feed the rglass in.

Print off one sheet of rglass. Print off all the metal that the rglass gave you, 9 or 10 sheets. Print off a stack of 19 glass.

Click on the glass with the rglass. Now you have 9 metal created from nothing, 20 rglass, and an autolathe ready to go again. Elapsed time: 5-10 seconds.

This brute-force fixes the metal glitch by making it so low value it's not worth attempting. If someone more experienced than me has a better idea of how to fix it, they're welcome to it.