In which two ways to get large quantities of metal for no cost are described fully and completely. (#6782)
* Metal b gone

Each T81 GL used to be worth 1.75 metal sheets when autolathed. No one ever takes them as they're single shot, so the 250 in the vendor became 430 metal sheets, or 8 and a half stacks.

They're now worth 10% of that, so yeah. No more shoveling half of the round's grenade launchers for near-unlimited metal.

* Duplication fix

Adds a decimal point to fix a duplication issue.

If you have an autolathe and a stack of 20 rglass, you can feed the rglass in.

Print off one sheet of rglass. Print off all the metal that the rglass gave you, 9 or 10 sheets. Print off a stack of 19 glass.

Click on the glass with the rglass. Now you have 9 metal created from nothing, 20 rglass, and an autolathe ready to go again. Elapsed time: 5-10 seconds.

This brute-force fixes the metal glitch by making it so low value it's not worth attempting. If someone more experienced than me has a better idea of how to fix it, they're welcome to it.