Add TS-34 Scopes (#7288)
Since the dawn of men we have been hunting. Step by step we improved our techniques to hunt animals. In the modern age many of the hunters use a Double Barrel shotgun, some hunters use scopes with it to shoot animals from a far distance. I am convinced that this addition will add marines a possibilty to use a new main weapon, instead of calling it a meme weaon.