Update ammo_datums.dm
After extensive testing, over a dozen or so matches since merge, as prae with neuro and trying combos, I feel it still lacks significant effectiveness in battle as marines are ignoring several neuro spits on direct hit as well as clouds and pushing forward, Prae's values has been changed, the cooldown of added spit delay reduced to 0 so the class only has the base cooldown from its castrum file. that being said the cost has increased to 200, giving prae a chance to scare marines, whilst having effectiveness.

I believe the idea of neuro spit is not meant to be used alone to kill marines, it should be used to assist the hive in slowing down pushes or counter attacking. right now it is not doing that job.