Adds in the attachables
It now can mount the hydro cannon, and the underbarrel shotgun and gl. Underbarrel flamer excluded for obvious reasons. Gonna have to check this with the back fuel tank, make sure it doesn't infinite-ammo these two for some ungodly reason.

Does not fill with water on spawn, as it is not the hydro cannon. Whoever coded that in kinda shitcoded it as more an alt-fire mode than an attachment, tbh.