Neurotoxic Spit Tweaks. (#7735)
* Neurotoxic Spit Tweaks.

* For some reason this changed?

* Update ammo_datums.dm

* Update ammo_datums.dm

After extensive testing, over a dozen or so matches since merge, as prae with neuro and trying combos, I feel it still lacks significant effectiveness in battle as marines are ignoring several neuro spits on direct hit as well as clouds and pushing forward, Prae's values has been changed, the cooldown of added spit delay reduced to 0 so the class only has the base cooldown from its castrum file. that being said the cost has increased to 200, giving prae a chance to scare marines, whilst having effectiveness.

I believe the idea of neuro spit is not meant to be used alone to kill marines, it should be used to assist the hive in slowing down pushes or counter attacking. right now it is not doing that job.

* Update ammo_datums.dm

fixed some typos

* Update ammo_datums.dm

typos

* Update ammo_datums.dm

Change cost for prae to allow 6 shots with the plasma pool in the prae buff pr that gives ancient 950

* Update ammo_datums.dm

Improved queen and sentinel overall values by 0.5 and decreased sent delay from 5 to 3.

* Update ammo_datums.dm

After more extensive testing, i nerfed Praes neuro damage down by 0.5 around the board except direct hit damage which recieved a flat -10. this goes better with the delay. the 450 typo was also removed to 45

* Update ammo_datums.dm

After extensive testing, these are the values I have decided suit the current need for Neuro spit to be best.

* Update ammo_datums.dm

* Update ammo_datums.dm

Reverted buff tweak to prae, lowered the plasma cost by 30 but also lowered the direct hit dmg to 40.

* Update ammo_datums.dm

Queen doesn't have neuro spit so the code is  removed

* Update ammo_datums.dm

never mind put back in

* Update ammo_datums.dm

Queen doesn't have neuro spit so this is goneeeeee

* Update ammo_datums.dm

fixing stuff