4x sleep healrate buff (#8278)
* Sleeping changes

First off, this uses regular ticks, not lifeticks, so there's that. You heal 20 times in the span that chems would heal you one, so let me just get that outta the way. The numbers are bigger than they look to be.

Adjusts health_ratio number to be -0.5 or greater, hence you'll heal if sleeping above crit. If you're crit, you're SOL, and probably getting eaten by a ravager or something.
Does fancy math to standardize the base healing amount to 0.25 per two seconds when in a bed with a bedsheet on it. Let's see... we want to divide that by 20 to get regular gameticks, so 0.0125, then divide *that* by 4 for 0.003125, and assign as needed to bed, table, floor, and bedsheet. Easy.

So now a standard bed with a bedsheet will heal a base amount of 0.25/2 second interval. Brute and burn stay as is, toxloss stays as is, staminaloss is multiplied by 100 (25, regular staminaheal is usually *higher* than that!), and cloneloss... cloneloss is adjusted to a base healrate of 0.2/ 2 seconds, at 100% hp.

Additionally, the cloneloss recovery rate is tied directly to how healthy you are via your health ratio. This means that if you go to sleep below 0 HP, or keep getting slashed in and out of crit and get unlucky on tickrate, you can end up with trace amounts of cloneloss.

Drunkenness left untouched.

* defines everything that's used in multiple places

* number change

4x given healrate