Adds RTG power cells
Adds smol and lorg RTG cells to the engi vendors. Three points for the smol one, 6 for the lorg one. Price of the lorg one is lower due to the difficulty of storing multiple.

Considered adding them to the SL vendor, but I think it's really too specialized for that.

Also considers adding the smol cell to the engi essential loadout, but it's really too specialized for even that.
yes i know the typepaths are wrong im changing it now reeee