Self-charging power cells (#8390)
* Adds two RTG power cells

The first one has 2000 maxcharge and regains 25 power every second. Fits in most things such as APC's or plasmacutters, but not weapons.
Power charge rate is enough to allow light use limited by max capacity. A manual charger will charge about 8 times faster, but this doesn't need your attention to charge it.

The second one has 5000 maxcharge and regains 50 power every second, but must be handcarried or kept stored inside equipment as it doesn't fit in baccpaccs or anything else (except powered thingies, obv). 
Power charge rate is enough for light use, and might allow powering a dedicated APC without external power, even with moderate loads.

And the handcharger, but I'mma remove that in another commit.

We have a lot of decent sprites sitting around also, neat.

* Prevents large RTG cells from fitting in the charger

Can't fit the giant cell into the smol charger and make a mockery of the weightclass system.

Also I might go back and make larg RTG cells a subtype of smol. That makes more sense.

* Adds RTG power cells

Adds smol and lorg RTG cells to the engi vendors. Three points for the smol one, 6 for the lorg one. Price of the lorg one is lower due to the difficulty of storing multiple.

Considered adding them to the SL vendor, but I think it's really too specialized for that.

Also considers adding the smol cell to the engi essential loadout, but it's really too specialized for even that.
yes i know the typepaths are wrong im changing it now reeee

* repaths RTG

Repaths small RTG to make more sense.

Considers making the larg RTG explode on destruction similar to OB ammo. Decided not to as of now.

* presses tab once

* I was wrong

first way was the right way to go

* repath

first way best way

* hits backspace exactly one time

* actual bugfixes

Fixes icon typepath with recharger cell
Fixes name of large recharger cell

* Phrasing

don't want the actual name of the item

* Price increases