Med changes
Changes kelotane and bicard to heal based off a curve. The first 5u heals 1/tick as normal, the next 5u heals a decreasing amount from 1.0 to 0.5/tick, and the next 5u and any more heal a flat 0.5/tick. The high-dose bonus (0.5 healrate if above 20u) remains unchanged. 

Also tricord got nerfed to a flat 0.5 healrate (and 0.25 toxheal) rather than 0.8... so it's 100% as good at total damageheal as a specialized med, not 160%.

Autoinjectors will still heal 25 points per inject, pills will now heal 56.25 points per pill (down from 75), and any chained pills will heal 37.5 points per pill.

That's the nerf. The buff is that addition of bicard or kelo now heals 20% of your brute/burn damage (plus five free points) *instantly* upon mob add, which can be substantial if you're hurt. The break-even point for pills to get equal or greater total heal than oldpills is ~68 health, while autoinjectors have the same performance and get the healrate changes as a straight bonus.

To prevent the obvious exploit of "well I'll just drink a buncha water and keep hittin' myself with a 1u bicakelo injector", ungas now get a *five* second stun when bicard and kelo leave your body, reduced by 0.5 seconds per tick it was in you. This can benefit ayys slightly, but in the worst-case scenario of you popping a standard 15u combat autoinjector *while* you're neurotoxined... 

...you still won't get stunned. 

It'll only get you if you have a microdose injector of less than 10u *and* are actively purging, or if you use an injector with less than 2u in it. Now, a *medic* running a microdose setup might be able to heal large quantities of damage rapidly on someone else at the expense of micromanagement and stunning the victim... but that's a different subject and one I'm more-or-less OK with.