Nerfs R-44 revolver
About The Pull Request
The current revolver meta is a very strong loadout where you use 20 or so revolvers all equipped with recoil compensators, red-dot sights and gyroscopic stabilizers which allows you to solo entire hives. I want to nerf the strength of the revolver so it stops a single marine from overpowering 4+ xenomorph players at once.

Why It's Good For The Game
"IDED/SALT PR" Jokes aside, the current state of the game is heavily in favor of the revolver. You've seen it yourself, and squashing the meta will be healthy for the game, in my honest opinion.  By nerfing it's fire rate, the total DPS is decreased and the amount of knockback per second is also decreased. The recoil and scatter should make it much harder to hit your shots which will make it much harder for a marine to easily spam away in the general direction of his enemies. The gun is still viable, it's just not trivial anymore to steamroll an entire hive with these.
Changelog
🆑
balance: Nerfs the accuracy, fire rate, recoil and scatter of the R-44 revolver.
/🆑