Brings our SDQL up to date (#13328)
* [READY]Ports yet another SDQL2 upgrade, refactors SDQL2 to a datum, adds SDQL2 options/more features (#42008)

Thanks to PJB3005
cl
experimental: SDQL2 has been refactored to a datum!
rscadd: A new SDQL2 panel has been added to admin tabs, for tracking, VVing, and halting SDQL2 queries.
rscadd: SDQL2 documentation is now available in SDQL_2.dm
rscadd: SDQL2 now has MAP added. MAP will cause the query to execute on whatever is specified in MAP, whether it's a variable or a procedure call (which will grab the return results), etc etc.
rscadd: SDQL2 now has a superuser mode, for uses outside of admin button pressing. This causes it to operate without admin protection wrapping.
rscadd: SDQL2 now supports options, including ignoring nulls in select or forcing it to operate in high priority mode, which lets it use 95% of the tick instead of obeying the Master Controller's tick limit. USE WITH CAUTION. Also includes a mode for blocking proccalls
rscadd: SDQL2 now supports TRUE/FALSE.
rscadd: To use options, append OPTIONS to the query. Available are "PRIORITY" = HIGH/NORMAL, "SELECT" = FORCE_NULLS/DISABLE or 0/FALSE, "PROCCALL" = ASYNC/BLOCKING.
/cl

Also displaytimetext is refactored.

* [READY] Adds selector lists to SDQL (#43383)

* [SDQL2] Adds selector arrays

* SDQL2 documentation spec updated slightly

* SDQL2 selector array docs + tweaks

If a single list is the only result of a MAP statement,
meaning objs.len == 1, AND it's a list, it will unwrap
the result from its containing list and return it directly.

The selector array logic keeps track of the state of the query
when it starts the first Search. It won't change the state back
to Execute until the top-level is reached, which sets a var
to keep track.

* SDQL2 MAP change from objs.len to length(objs)

* [SDQL2] Reverts MAP list unwrap

* [SDQL] Move (expr) from expr, unary_expr to var

This should result in the same functionality but instead of expressions and unary_expressions processing sub expressions, variables do
This means that expression->value->variable is now completely recursable

* [SDQL2] Update selector list documentation

* Adds sequential mode to SDQL

* Fix sequential SDQL being default (#44476)

* SDQL2 dynamically gets subsystems (#48335)

As opposed to a static switch statement
This means all present and future subsystems are supported
without needing to add them to the switch

* Allows calling procs in SDQL update queries without storing the return anywhere. (#48690)

* Adds everything that's needed for SDQL varedit macros

* Removes unused proc

* Allows calling procs in update queries without setting the return anywhere.

* Readds a line

* Removes file wrapper

* Unicode awareness Part 2 -- copytext()

* Fixes an SQDL2 runtime error (#49327)

If you tried to call SQDL with " at the end, it would throw eg WHERE var="something"

* Misc spellchecks (tg#51134)

* Missed an s

* Linters

* Undo changed spans + deleted wrappers

* Turn these back to length

* Few more

---------

Co-authored-by: JJRcop <jrubcop@gmail.com>
Co-authored-by: AnturK <AnturK@users.noreply.github.com>
Co-authored-by: nemvar <47324920+nemvar@users.noreply.github.com>
Co-authored-by: MrPerson <spamtaffic@gmail.com>