Adds file cycling for icon2base64 (#76395)
I was going through condensed runtime stastics earlier today and noticed
that we're getting thousands of runtimes from icon2base64 trying to
write into a read only savefile; which can only happen if that savefile
is opened more than once, (or is Locked)

This is likely due to some logic somewhere holding onto a reference it
really shouldn't; but also is almost certainly a byond issue since the
proc also clears the reference immediately after doing its work.

In order to combat this I've added a simple file cycling setup for
icon2base64 so that it can hopefully not run into that issue again and
by the time it gets back to a previous savefile byond has cleaned up its
lock.

See the top of
https://tgstation13.download/parsed-logs/manuel/data/logs/2023/06/28/round-209229/runtime.condensed.txt